refactor: vket/views.py を機能別に分割
1482行の vket/views.py を8つのモジュールに分割し、
__init__.py で re-export して後方互換を維持。

- helpers.py: ヘルパー関数群、定数、Slot dataclass
- public.py: CollaborationListView, CollaborationDetailView
- apply.py: ApplyView
- manage.py: ManageView, ManageParticipationUpdateView, ManageScheduleView
- notice.py: NoticeListView, ManageNoticeListView, ManageNoticeCreateView, ManageNoticeUpdateView, AckNoticeView
- status.py: VketStatusRedirectView, StageRegisterView, ParticipationStatusView
- presentation.py: PresentationDeleteView, ManagePresentationDeleteView
- publish.py: ManagePublishView

Closes #199
